backend/requirements.txt
fastapi==0.110.3
uvicorn==0.29.0
docker-py
python-dotenv==1.0.1
